removing unneeded hooks
