+++ src/components/color-block.tsx
@@ -1,24 +1,22 @@
 import chroma from "chroma-js";
+import { useCallback } from "react";
 import { FiCopy } from "react-icons/fi";
-import { useCopy } from "./useCopy";
 
-const ColorBlockItem = ({
-  hex,
-  name,
-  color,
-}: {
+interface IBlockItemProps {
   hex: string;
   name: string;
   color: string;
-}) => {
-  const [, handleCopy] = useCopy();
+  onCopy(value: string): void;
+}
+
+const ColorBlockItem = ({ name, color, onCopy }: IBlockItemProps) => {
   return (
     <div className="root">
       <div>
         <dt>{name}</dt>
         <dd>{color}</dd>
       </div>
-      <button onClick={() => handleCopy(color)}>
+      <button onClick={() => onCopy(color)}>
         <FiCopy />
       </button>
 
@@ -57,37 +55,64 @@ const ColorBlockItem = ({
   );
 };
 
-export const ColorBlock = ({ color }: { color: string }) => (
-  <figure>
-    <object />
-    <figcaption>
-      <dl>
-        <ColorBlockItem hex={color} name="Hex" color={color.toUpperCase()} />
-        <ColorBlockItem
-          hex={color}
-          name="HSL"
-          color={chroma(color).css("hsl")}
-        />
-        <ColorBlockItem hex={color} name="RGB" color={chroma(color).css()} />
-      </dl>
-    </figcaption>
+interface IColorBlockProps {
+  color: string;
+  hsl: string;
+  rgb: string;
+}
 
-    <style jsx>{`
-      figure {
-        border-radius: 0.25rem;
-        overflow: hidden;
-        box-shadow: 0 1px 3px 0 rgb(0 0 0 / 10%), 0 1px 2px 0 rgb(0 0 0 / 6%);
-        margin-bottom: 2rem;
-      }
+export const ColorBlock = ({ color }: IColorBlockProps) => {
+  const handleCopy = useCallback(async (value: string) => {
+    try {
+      await navigator.clipboard.writeText(value);
+    } catch (err: any) {
+      console.log(`Failed to copy: ${err.message}`);
+    }
+  }, []);
 
-      object {
-        aspect-ratio: 16 / 9;
-        background: ${color};
-        width: 100%;
-      }
-      dl {
-        padding: 0.5rem;
-      }
-    `}</style>
-  </figure>
-);
+  return (
+    <figure>
+      <object />
+      <figcaption>
+        <dl>
+          <ColorBlockItem
+            onCopy={handleCopy}
+            hex={color}
+            name="Hex"
+            color={color.toUpperCase()}
+          />
+          <ColorBlockItem
+            onCopy={handleCopy}
+            hex={color}
+            name="HSL"
+            color={chroma(color).css("hsl")}
+          />
+          <ColorBlockItem
+            onCopy={handleCopy}
+            hex={color}
+            name="RGB"
+            color={chroma(color).css()}
+          />
+        </dl>
+      </figcaption>
+
+      <style jsx>{`
+        figure {
+          border-radius: 0.25rem;
+          overflow: hidden;
+          box-shadow: 0 1px 3px 0 rgb(0 0 0 / 10%), 0 1px 2px 0 rgb(0 0 0 / 6%);
+          margin-bottom: 2rem;
+        }
+
+        object {
+          aspect-ratio: 16 / 9;
+          background: ${color};
+          width: 100%;
+        }
+        dl {
+          padding: 0.5rem;
+        }
+      `}</style>
+    </figure>
+  );
+};
